# specs/003-develop-ui/contracts/bff-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: DevOps Career Agent Browser BFF API
  version: 1.0.0
  description: >-
    Canonical UI-to-BFF contract. Browser-facing payloads are presentation-safe
    camelCase DTOs. The BFF validates this boundary, maps to the private core
    contract, and never exposes OAuth tokens, authority claims, answer keys, or
    another employee's identifiers.
servers:
  - url: /bff/v1
security:
  - sessionCookie: []
paths:
...
  /learning-sessions:
    get:
      operationId: listBrowserLearningSessions
      parameters:
        - $ref: "#/components/parameters/RoadmapIdQuery"
      responses:
        "200":
          description: Employee-owned learning sessions for a roadmap.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningSession"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /learning-sessions/{id}:
    get:
      operationId: getBrowserLearningSession
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Persisted learning session and current cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSession"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /learning-sessions/{id}/start:
    post:
      operationId: startBrowserLearningSession
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: Started or resumed learning session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSession"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /learning-sessions/{id}/steps/{stepId}/completion:
    put:
      operationId: completeBrowserLearningStep
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: stepId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: Persisted step completion and updated session cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSession"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /learning-sessions/{id}/review-attempts:
    get:
      operationId: listBrowserReviewAttempts
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Safe persisted review-attempt history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewAttempt"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
    post:
      operationId: createBrowserReviewAttempt
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: New or established in-progress review attempt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InProgressReviewAttempt"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /review-attempts/{id}:
    get:
      operationId: getBrowserReviewAttempt
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Safe persisted review attempt and submitted answers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAttempt"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /review-attempts/{id}/answers/{questionId}:
    put:
      operationId: answerBrowserReviewQuestion
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAnswerRequest"
      responses:
        "200":
          description: Feedback for the answered question only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAnswerFeedback"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /review-attempts/{id}/submit:
    post:
      operationId: submitBrowserReviewAttempt
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: Immutable score, completion status, and one next action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResult"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /lab-references/{id}/reports:
    post:
      operationId: reportBrowserLabReference
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabReportRequest"
      responses:
        "202":
          description: Report accepted without changing learning completion.
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /learning/next-action:
    get:
      operationId: getBrowserNextLearningAction
      parameters:
        - $ref: "#/components/parameters/RoadmapIdQuery"
      responses:
        "200":
          description: Exactly one prioritized next action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextAction"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-learning_session
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    RoadmapIdQuery:
      name: roadmapId
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    UiContractVersion:
      name: X-UI-Contract-Version
      in: header
      required: true
      schema:
        type: string
        pattern: "^1\\.[0-9]+\\.[0-9]+$"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
...
  responses:
    SessionNotActive:
      description: Missing, expired, revoked, or otherwise invalid BFF session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    Forbidden:
      description: Session, Origin, CSRF, lifecycle, or owner policy denied access.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Owner-scoped resource does not exist or is not visible.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Contract-version, idempotency, or state conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationFailed:
      description: Request schema, field, path, or query validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RateLimited:
      description: Browser actor or required dependency is temporarily throttled.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ServiceUnavailable:
      description: BFF, core, identity, session, or persistence dependency unavailable.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    GatewayTimeout:
      description: A bounded downstream operation timed out with safe retry semantics.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    ReviewAnswerRequest:
      type: object
      additionalProperties: false
      required: [answerKey]
      properties:
        answerKey:
          type: string
          minLength: 1
    LabReportRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          enum: [unavailable, unsuitable, cost_mismatch, other]
        comment:
          type: [string, "null"]
          maxLength: 1000
...
    LabReference:
      type: object
      additionalProperties: false
      required:
        - id
        - provider
        - objective
        - prerequisites
        - estimatedMinutes
        - costStatus
        - availabilityStatus
        - destinationUrl
      properties:
        id:
          type: string
        provider:
          type: string
        objective:
          type: string
        prerequisites:
          type: array
          items:
            type: string
        estimatedMinutes:
          type: integer
          minimum: 1
        costStatus:
          type: string
          enum: [free, paid, subscription, unknown]
        availabilityStatus:
          type: string
          enum: [active, unavailable, retired]
        destinationUrl:
          type: string
          format: uri
          pattern: "^https://"
    LearningSession:
      type: object
      additionalProperties: false
      required:
        - id
        - roadmapId
        - title
        - objective
        - estimatedMinutes
        - status
        - currentStepOrdinal
        - steps
        - retirement
      properties:
        id:
          type: string
        roadmapId:
          type: string
        contentVersion:
          type: string
        questionVersion:
          type: string
        labReferenceVersions:
          type: array
          items:
            type: string
        title:
          type: string
        objective:
          type: string
        estimatedMinutes:
          type: integer
          minimum: 20
          maximum: 30
        status:
          type: string
          enum: [not_started, in_progress, retry_required, completed]
        currentStepOrdinal:
          type: integer
          minimum: 0
        steps:
          type: array
          items:
            $ref: "#/components/schemas/LearningStep"
        retirement:
          type: object
          additionalProperties: false
          required: [status, securityCritical]
          properties:
            status:
              type: string
              enum: [draft, published, retired]
            securityCritical:
              type: boolean
            resumeUntil:
              type: [string, "null"]
              format: date-time
            replacementContentId:
              type: [string, "null"]
            replacementContentVersion:
              type: [string, "null"]
    LearningStep:
      type: object
      additionalProperties: false
      required: [id, ordinal, stepType, title, status]
      properties:
        id:
          type: string
        ordinal:
          type: integer
          minimum: 0
        stepType:
          type: string
          enum: [reading, activity, lab, review]
        title:
          type: string
          minLength: 1
        status:
          type: string
          enum: [pending, completed]
        labReference:
          oneOf:
            - $ref: "#/components/schemas/LabReference"
            - type: "null"
    ReviewAttempt:
      oneOf:
        - $ref: "#/components/schemas/InProgressReviewAttempt"
        - $ref: "#/components/schemas/SubmittedReviewAttempt"
      discriminator:
        propertyName: status
        mapping:
          in_progress: "#/components/schemas/InProgressReviewAttempt"
          submitted: "#/components/schemas/SubmittedReviewAttempt"
    InProgressReviewAttempt:
      type: object
      additionalProperties: false
      required:
        - id
        - attemptNumber
        - questionSnapshotVersion
        - status
        - questions
        - answers
        - startedAt
        - scorePercent
        - passed
        - submittedAt
      properties:
        id:
          type: string
        attemptNumber:
          type: integer
          minimum: 1
        questionSnapshotVersion:
          type: string
        status:
          type: string
          const: in_progress
        questions:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: "#/components/schemas/ReviewQuestion"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PersistedReviewAnswer"
        startedAt:
          type: string
          format: date-time
        scorePercent:
          type: "null"
        passed:
          type: "null"
        submittedAt:
          type: "null"
        latest:
          type: boolean
        highest:
          type: boolean
    SubmittedReviewAttempt:
      type: object
      additionalProperties: false
      required:
        - id
        - attemptNumber
        - questionSnapshotVersion
        - status
        - questions
        - answers
        - startedAt
        - scorePercent
        - passed
        - submittedAt
      properties:
        id:
          type: string
        attemptNumber:
          type: integer
          minimum: 1
        questionSnapshotVersion:
          type: string
        status:
          type: string
          const: submitted
        questions:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: "#/components/schemas/ReviewQuestion"
        answers:
          type: array
          minItems: 3
          maxItems: 5
          items:
            $ref: "#/components/schemas/PersistedReviewAnswer"
        startedAt:
          type: string
          format: date-time
        scorePercent:
          type: number
          minimum: 0
          maximum: 100
        passed:
          type: boolean
        submittedAt:
          type: string
          format: date-time
        latest:
          type: boolean
        highest:
          type: boolean
    ReviewQuestion:
      type: object
      additionalProperties: false
      required: [id, ordinal, prompt, choices]
      properties:
        id:
          type: string
        ordinal:
          type: integer
          minimum: 0
        prompt:
          type: string
          minLength: 1
        choices:
          type: object
          minProperties: 2
          additionalProperties:
            type: string
          description: Correct keys are excluded.
    PersistedReviewAnswer:
      type: object
      additionalProperties: false
      description: >-
        Safe restoration DTO for an answered question; no authoritative correct
        key for any unanswered question is exposed.
      required:
        - questionId
        - submittedAnswerKey
        - correct
        - explanation
        - answeredAt
      properties:
        questionId:
          type: string
        submittedAnswerKey:
          type: string
        correct:
          type: boolean
        explanation:
          type: string
          minLength: 1
        answeredAt:
          type: string
          format: date-time
    ReviewAnswerFeedback:
      type: object
      additionalProperties: false
      required:
        - questionId
        - submittedAnswerKey
        - correct
        - explanation
        - answeredAt
      properties:
        questionId:
          type: string
        submittedAnswerKey:
          type: string
        correct:
          type: boolean
        explanation:
          type: string
          minLength: 1
        answeredAt:
          type: string
          format: date-time
    ReviewResult:
      type: object
      additionalProperties: false
      required: [attemptId, scorePercent, passed, sessionStatus, missedQuestionIds, nextAction]
      properties:
        attemptId:
          type: string
        scorePercent:
          type: number
          minimum: 0
          maximum: 100
        passed:
          type: boolean
        sessionStatus:
          type: string
          enum: [retry_required, completed]
        missedQuestionIds:
          type: array
          items:
            type: string
        nextAction:
          $ref: "#/components/schemas/NextAction"
...
    NextAction:
      type: object
      additionalProperties: false
      required: [kind, title, reason, target]
      properties:
        kind:
          type: string
          enum: [retry_material, resume_session, continue_milestone, review_roadmap]
        title:
          type: string
          minLength: 1
        reason:
          type: string
          minLength: 1
        target:
          type: string
          minLength: 1
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId, retryable, fieldErrors]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: [string, "null"]
        code:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, messages]
            properties:
              field:
                type: string
              messages:
                type: array
                minItems: 1
                items:
                  type: string